pydice 16 renamed roll_result_operators package.
