+++ pydice/dice_string/operators.py
@@ -1,7 +1,7 @@
 from abc import abstractmethod, ABC
 from dataclasses import dataclass
 
-from pydice.roll_result_modifiers.roll_result_builder import RollResultBuilder
+from pydice.roll_result_operators.roll_result_builder import RollResultBuilder
 
 
 @dataclass
