+++ pydice/roll_result_operators/roll_result_builder.py
@@ -1,8 +1,8 @@
 from pydice.die import Dice
 from pydice.roll_result import RollResult, DiceRollResult
-from pydice.roll_result_modifiers.roll_result_decorators import AddToRollResultDecorator, SubtractFromRollResultDecorator, \
+from pydice.roll_result_operators.roll_result_decorators import AddToRollResultDecorator, SubtractFromRollResultDecorator, \
     MultiplyRollResultDecorator, DivideByRollResultDecorator, ExplodeDiceForTargetDecorator
-from pydice.roll_result_modifiers.counter_roll_result_decorator import CountValuesEqualToDecorator, CountValuesGreaterThanDecorator, \
+from pydice.roll_result_operators.counter_roll_result_decorator import CountValuesEqualToDecorator, CountValuesGreaterThanDecorator, \
     CountValuesLessThanDecorator, CountValuesNotEqualToDecorator
