+++ pydice/roll_result_operators/roll_result_decorators.py
