+++ tests/dice_strings_tests/test_default_dice_string_parser.py
@@ -7,9 +7,9 @@ from pydice.die import Dice, Die
 from pydice.dice_string.operators import AddOperator, SubtractOperator, MultiplyOperator, DivideOperator, EqualToOperator, \
     GreaterThanEqualToOperator, GreaterThanOperator, LessThanOperator, LessThanEqualToOperator, NotEqualToOperator
 from pydice.roll_result import DiceRollResult
-from pydice.roll_result_modifiers.roll_result_decorators import AddToRollResultDecorator, SubtractFromRollResultDecorator, \
+from pydice.roll_result_operators.roll_result_decorators import AddToRollResultDecorator, SubtractFromRollResultDecorator, \
     MultiplyRollResultDecorator, DivideByRollResultDecorator
-from pydice.roll_result_modifiers.counter_roll_result_decorator import CountValuesEqualToDecorator, CountValuesGreaterThanDecorator, \
+from pydice.roll_result_operators.counter_roll_result_decorator import CountValuesEqualToDecorator, CountValuesGreaterThanDecorator, \
     CountValuesLessThanDecorator, CountValuesNotEqualToDecorator
 
 dice_results = [9, 10, 6, 7, 6, 1, 2, 4, 8, 3]
