+++ tests/roll_result_modifiers_tests/test_combined_mathematical_roll_result_decorators.py
@@ -1,7 +1,7 @@
 import unittest
 from unittest.mock import Mock, PropertyMock
 
-from pydice.roll_result_modifiers.roll_result_decorators import AddToRollResultDecorator, SubtractFromRollResultDecorator, \
+from pydice.roll_result_operators.roll_result_decorators import AddToRollResultDecorator, SubtractFromRollResultDecorator, \
     MultiplyRollResultDecorator, DivideByRollResultDecorator
