+++ tests/roll_result_modifiers_tests/test_divide_by_roll_result_decorator.py
@@ -1,7 +1,7 @@
 import unittest
 from unittest.mock import Mock, PropertyMock
 
-from pydice.roll_result_modifiers.roll_result_decorators import DivideByRollResultDecorator
+from pydice.roll_result_operators.roll_result_decorators import DivideByRollResultDecorator
 
 
 class DivideByRollResultDecoratorTests(unittest.TestCase):
