+++ tests/roll_result_modifiers_tests/test_explode_dice_for_target_decorator.py
@@ -3,7 +3,7 @@ from unittest.mock import MagicMock
 
 from pydice.die import Die
 from pydice.roll_result import RollResult
-from pydice.roll_result_modifiers.roll_result_decorators import ExplodeDiceForTargetDecorator
+from pydice.roll_result_operators.roll_result_decorators import ExplodeDiceForTargetDecorator
 
 
 class ExplodeDiceForTargetDecoratorTests(unittest.TestCase):
